glfw_sys

Function glfwGetGammaRamp

pub unsafe extern "C" fn glfwGetGammaRamp(
    monitor: *mut GLFWmonitor,
) -> *const GLFWgammaramp

@brief Returns the current gamma ramp for the specified monitor.

This function returns the current gamma ramp of the specified monitor.

@param[in] monitor The monitor to query. @return The current gamma ramp, or NULL if an [error](@ref error_handling) occurred.

@errors Possible errors include @ref GLFW_NOT_INITIALIZED, @ref GLFW_PLATFORM_ERROR and @ref GLFW_FEATURE_UNAVAILABLE (see remarks).

@remark @wayland Gamma handling is a privileged protocol, this function will thus never be implemented and emits @ref GLFW_FEATURE_UNAVAILABLE while returning NULL.

@pointer_lifetime The returned structure and its arrays are allocated and freed by GLFW. You should not free them yourself. They are valid until the specified monitor is disconnected, this function is called again for that monitor or the library is terminated.

@thread_safety This function must only be called from the main thread.

@sa @ref monitor_gamma

@since Added in version 3.0.

@ingroup monitor