glfw_sys

Function glfwGetPhysicalDevicePresentationSupport

pub unsafe extern "C" fn glfwGetPhysicalDevicePresentationSupport(
    instance: VkInstance,
    device: VkPhysicalDevice,
    queuefamily: u32,
) -> c_int

@brief Returns whether the specified queue family can present images.

This function returns whether the specified queue family of the specified physical device supports presentation to the platform GLFW was built for.

If Vulkan or the required window surface creation instance extensions are not available on the machine, or if the specified instance was not created with the required extensions, this function returns GLFW_FALSE and generates a @ref GLFW_API_UNAVAILABLE error. Call @ref glfwVulkanSupported to check whether Vulkan is at least minimally available and @ref glfwGetRequiredInstanceExtensions to check what instance extensions are required.

@param[in] instance The instance that the physical device belongs to. @param[in] device The physical device that the queue family belongs to. @param[in] queuefamily The index of the queue family to query. @return GLFW_TRUE if the queue family supports presentation, or GLFW_FALSE otherwise.

@errors Possible errors include @ref GLFW_NOT_INITIALIZED, @ref GLFW_API_UNAVAILABLE and @ref GLFW_PLATFORM_ERROR.

@remark @macos This function currently always returns GLFW_TRUE, as the VK_MVK_macos_surface and VK_EXT_metal_surface extensions do not provide a vkGetPhysicalDevice*PresentationSupport type function.

@thread_safety This function may be called from any thread. For synchronization details of Vulkan objects, see the Vulkan specification.

@sa @ref vulkan_present

@since Added in version 3.2.

@ingroup vulkan