glfw_sys

Function glfwGetVersionString

pub unsafe extern "C" fn glfwGetVersionString() -> *const c_char

@brief Returns a string describing the compile-time configuration.

This function returns the compile-time generated [version string](@ref intro_version_string) of the GLFW library binary. It describes the version, platforms, compiler and any platform or operating system specific compile-time options. It should not be confused with the OpenGL or OpenGL ES version string, queried with glGetString.

Do not use the version string to parse the GLFW library version. The @ref glfwGetVersion function provides the version of the running library binary in numerical format.

Do not use the version string to parse what platforms are supported. The @ref glfwPlatformSupported function lets you query platform support.

@return The ASCII encoded GLFW version string.

@errors None.

@remark This function may be called before @ref glfwInit.

@pointer_lifetime The returned string is static and compile-time generated.

@thread_safety This function may be called from any thread.

@sa @ref intro_version @sa @ref glfwGetVersion

@since Added in version 3.0.

@ingroup init