glfw_sys

Function glfwRawMouseMotionSupported

pub unsafe extern "C" fn glfwRawMouseMotionSupported() -> c_int

@brief Returns whether raw mouse motion is supported.

This function returns whether raw mouse motion is supported on the current system. This status does not change after GLFW has been initialized so you only need to check this once. If you attempt to enable raw motion on a system that does not support it, @ref GLFW_PLATFORM_ERROR will be emitted.

Raw mouse motion is closer to the actual motion of the mouse across a surface. It is not affected by the scaling and acceleration applied to the motion of the desktop cursor. That processing is suitable for a cursor while raw motion is better for controlling for example a 3D camera. Because of this, raw mouse motion is only provided when the cursor is disabled.

@return GLFW_TRUE if raw mouse motion is supported on the current machine, or GLFW_FALSE otherwise.

@errors Possible errors include @ref GLFW_NOT_INITIALIZED.

@thread_safety This function must only be called from the main thread.

@sa @ref raw_mouse_motion @sa @ref glfwSetInputMode

@since Added in version 3.3.

@ingroup input