glfw_sys

Function glfwSetClipboardString

pub unsafe extern "C" fn glfwSetClipboardString(
    window: *mut GLFWwindow,
    string: *const c_char,
)

@brief Sets the clipboard to the specified string.

This function sets the system clipboard to the specified, UTF-8 encoded string.

@param[in] window Deprecated. Any valid window or NULL. @param[in] string A UTF-8 encoded string.

@errors Possible errors include @ref GLFW_NOT_INITIALIZED and @ref GLFW_PLATFORM_ERROR.

@remark @win32 The clipboard on Windows has a single global lock for reading and writing. GLFW tries to acquire it a few times, which is almost always enough. If it cannot acquire the lock then this function emits @ref GLFW_PLATFORM_ERROR and returns. It is safe to try this multiple times.

@pointer_lifetime The specified string is copied before this function returns.

@thread_safety This function must only be called from the main thread.

@sa @ref clipboard @sa @ref glfwGetClipboardString

@since Added in version 3.0.

@ingroup input