glfw_sys

Function glfwSetCursorPos

pub unsafe extern "C" fn glfwSetCursorPos(
    window: *mut GLFWwindow,
    xpos: f64,
    ypos: f64,
)

@brief Sets the position of the cursor, relative to the content area of the window.

This function sets the position, in screen coordinates, of the cursor relative to the upper-left corner of the content area of the specified window. The window must have input focus. If the window does not have input focus when this function is called, it fails silently.

Do not use this function to implement things like camera controls. GLFW already provides the GLFW_CURSOR_DISABLED cursor mode that hides the cursor, transparently re-centers it and provides unconstrained cursor motion. See @ref glfwSetInputMode for more information.

If the cursor mode is GLFW_CURSOR_DISABLED then the cursor position is unconstrained and limited only by the minimum and maximum values of a double.

@param[in] window The desired window. @param[in] xpos The desired x-coordinate, relative to the left edge of the content area. @param[in] ypos The desired y-coordinate, relative to the top edge of the content area.

@errors Possible errors include @ref GLFW_NOT_INITIALIZED, @ref GLFW_PLATFORM_ERROR and @ref GLFW_FEATURE_UNAVAILABLE (see remarks).

@remark @wayland This function will only work when the cursor mode is GLFW_CURSOR_DISABLED, otherwise it will emit @ref GLFW_FEATURE_UNAVAILABLE.

@thread_safety This function must only be called from the main thread.

@sa @ref cursor_pos @sa @ref glfwGetCursorPos

@since Added in version 3.0. Replaces glfwSetMousePos.

@ingroup input