glfw_sys

Function glfwSetWindowAttrib

pub unsafe extern "C" fn glfwSetWindowAttrib(
    window: *mut GLFWwindow,
    attrib: c_int,
    value: c_int,
)

@brief Sets an attribute of the specified window.

This function sets the value of an attribute of the specified window.

The supported attributes are GLFW_DECORATED(@ref GLFW_DECORATED_attrib), GLFW_RESIZABLE(@ref GLFW_RESIZABLE_attrib), GLFW_FLOATING(@ref GLFW_FLOATING_attrib), GLFW_AUTO_ICONIFY(@ref GLFW_AUTO_ICONIFY_attrib) and GLFW_FOCUS_ON_SHOW(@ref GLFW_FOCUS_ON_SHOW_attrib). GLFW_MOUSE_PASSTHROUGH(@ref GLFW_MOUSE_PASSTHROUGH_attrib)

Some of these attributes are ignored for full screen windows. The new value will take effect if the window is later made windowed.

Some of these attributes are ignored for windowed mode windows. The new value will take effect if the window is later made full screen.

@param[in] window The window to set the attribute for. @param[in] attrib A supported window attribute. @param[in] value GLFW_TRUE or GLFW_FALSE.

@errors Possible errors include @ref GLFW_NOT_INITIALIZED, @ref GLFW_INVALID_ENUM, @ref GLFW_INVALID_VALUE, @ref GLFW_PLATFORM_ERROR and @ref GLFW_FEATURE_UNAVAILABLE (see remarks).

@remark Calling @ref glfwGetWindowAttrib will always return the latest value, even if that value is ignored by the current mode of the window.

@remark @wayland The GLFW_FLOATING(@ref GLFW_FLOATING_attrib) window attribute is not supported. Setting this will emit @ref GLFW_FEATURE_UNAVAILABLE.

@thread_safety This function must only be called from the main thread.

@sa @ref window_attribs @sa @ref glfwGetWindowAttrib

@since Added in version 3.3.

@ingroup window