glfw_sys

Function glfwSetWindowSize

pub unsafe extern "C" fn glfwSetWindowSize(
    window: *mut GLFWwindow,
    width: c_int,
    height: c_int,
)

@brief Sets the size of the content area of the specified window.

This function sets the size, in screen coordinates, of the content area of the specified window.

For full screen windows, this function updates the resolution of its desired video mode and switches to the video mode closest to it, without affecting the window’s context. As the context is unaffected, the bit depths of the framebuffer remain unchanged.

If you wish to update the refresh rate of the desired video mode in addition to its resolution, see @ref glfwSetWindowMonitor.

The window manager may put limits on what sizes are allowed. GLFW cannot and should not override these limits.

@param[in] window The window to resize. @param[in] width The desired width, in screen coordinates, of the window content area. @param[in] height The desired height, in screen coordinates, of the window content area.

@errors Possible errors include @ref GLFW_NOT_INITIALIZED and @ref GLFW_PLATFORM_ERROR.

@thread_safety This function must only be called from the main thread.

@sa @ref window_size @sa @ref glfwGetWindowSize @sa @ref glfwSetWindowMonitor

@since Added in version 1.0. @glfw3 Added window handle parameter.

@ingroup window