glfw_sys

Function glfwTerminate

pub unsafe extern "C" fn glfwTerminate()

@brief Terminates the GLFW library.

This function destroys all remaining windows and cursors, restores any modified gamma ramps and frees any other allocated resources. Once this function is called, you must again call @ref glfwInit successfully before you will be able to use most GLFW functions.

If GLFW has been successfully initialized, this function should be called before the application exits. If initialization fails, there is no need to call this function, as it is called by @ref glfwInit before it returns failure.

This function has no effect if GLFW is not initialized.

@errors Possible errors include @ref GLFW_PLATFORM_ERROR.

@remark This function may be called before @ref glfwInit.

@warning The contexts of any remaining windows must not be current on any other thread when this function is called.

@reentrancy This function must not be called from a callback.

@thread_safety This function must only be called from the main thread.

@sa @ref intro_init @sa @ref glfwInit

@since Added in version 1.0.

@ingroup init